+++ src/main/java/timify/com/TempController.java
@@ -1,11 +1,11 @@
-package timify.com.controller;
+package timify.com;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
-import timify.com.apiPayload.code.status.ErrorStatus;
-import timify.com.apiPayload.exception.handler.TempHandler;
+import timify.com.common.apiPayload.code.status.ErrorStatus;
+import timify.com.common.apiPayload.exception.handler.TempHandler;
 
 @RestController
 @RequiredArgsConstructor
@@ -13,7 +13,7 @@ import timify.com.apiPayload.exception.handler.TempHandler;
 public class TempController {
 
     @GetMapping("/health")
-    public String healthCheck(){
+    public String healthCheck() {
         return "Hello";
     }
