+++ src/main/java/timify/com/auth/AuthController.java
@@ -1,0 +1,15 @@
+package timify.com.auth;
+
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@Slf4j
+@RequestMapping("/v1/auth")
+public class AuthController {
+
+}
