+++ src/main/java/timify/com/common/apiPayload/code/BaseErrorCode.java
@@ -1,4 +1,4 @@
-package timify.com.apiPayload.code;
+package timify.com.common.apiPayload.code;
 
 public interface BaseErrorCode {
     ErrorReasonDTO getReason();
