+++ src/main/java/timify/com/common/apiPayload/code/ErrorReasonDTO.java
@@ -1,4 +1,4 @@
-package timify.com.apiPayload.code;
+package timify.com.common.apiPayload.code;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,9 +6,9 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 @Builder
-public class ReasonDTO {
+public class ErrorReasonDTO {
 
-    private HttpStatus httpStatus;
+    private final HttpStatus httpStatus;
 
     private final boolean isSuccess;
     private final String code;
