+++ src/main/java/timify/com/common/apiPayload/exception/ExceptionAdvice.java
@@ -1,4 +1,4 @@
-package timify.com.apiPayload.exception;
+package timify.com.common.apiPayload.exception;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.ConstraintViolationException;
@@ -14,9 +14,9 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.ServletWebRequest;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-import timify.com.apiPayload.ApiResponse;
-import timify.com.apiPayload.code.ErrorReasonDTO;
-import timify.com.apiPayload.code.status.ErrorStatus;
+import timify.com.common.apiPayload.ApiResponse;
+import timify.com.common.apiPayload.code.ErrorReasonDTO;
+import timify.com.common.apiPayload.code.status.ErrorStatus;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -49,7 +49,7 @@ public class ExceptionAdvice extends ResponseEntityExceptionHandler {
                     errors.merge(fieldName, errorMessage, (existingErrorMessage, newErrorMessage) -> existingErrorMessage + ", " + newErrorMessage);
                 });
 
-        return handleExceptionInternalArgs(e,HttpHeaders.EMPTY,ErrorStatus.valueOf("_BAD_REQUEST"),request,errors);
+        return handleExceptionInternalArgs(e, HttpHeaders.EMPTY, ErrorStatus.valueOf("_BAD_REQUEST"), request, errors);
     }
 
     @org.springframework.web.bind.annotation.ExceptionHandler
