+++ src/main/java/timify/com/common/apiPayload/exception/GeneralException.java
@@ -1,9 +1,9 @@
-package timify.com.apiPayload.exception;
+package timify.com.common.apiPayload.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import timify.com.apiPayload.code.BaseErrorCode;
-import timify.com.apiPayload.code.ErrorReasonDTO;
+import timify.com.common.apiPayload.code.BaseErrorCode;
+import timify.com.common.apiPayload.code.ErrorReasonDTO;
 
 @Getter
 @AllArgsConstructor
