+++ src/main/java/timify/com/common/apiPayload/exception/handler/TempHandler.java
@@ -1,0 +1,10 @@
+package timify.com.common.apiPayload.exception.handler;
+
+import timify.com.common.apiPayload.code.BaseErrorCode;
+import timify.com.common.apiPayload.exception.GeneralException;
+
+public class TempHandler extends GeneralException {
+    public TempHandler(BaseErrorCode errorCode) {
+        super(errorCode);
+    }
+}
