+++ src/main/java/timify/com/domain/StudyPlace.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.*;
 import timify.com.domain.common.BaseDateTimeEntity;
 import timify.com.domain.enums.CategoryStatus;
+import timify.com.member.domain.Member;
 
 @Entity
 @Getter
