+++ src/main/java/timify/com/domain/StudyTime.java
@@ -3,6 +3,7 @@ package timify.com.domain;
 import jakarta.persistence.*;
 import lombok.*;
 import timify.com.domain.common.BaseDateTimeEntity;
+import timify.com.member.domain.Member;
 
 import java.time.LocalDateTime;
