--- src/main/java/timify/com/domain/enums/LoginType.java
@@ -1,5 +1,0 @@
-package timify.com.domain.enums;
-
-public enum LoginType {
-    KAKAO,APPLE
-}
