+++ src/main/java/timify/com/member/MemberController.java
@@ -1,0 +1,13 @@
+package timify.com.member;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@Slf4j
+@RequestMapping("/v1/member")
+public class MemberController {
+}
