+++ src/main/java/timify/com/member/domain/Gender.java
@@ -1,4 +1,4 @@
-package timify.com.domain.enums;
+package timify.com.member.domain;
 
 public enum Gender {
     MALE, FEMALE, NONE
