+++ src/main/java/timify/com/member/domain/LoginType.java
@@ -1,0 +1,5 @@
+package timify.com.member.domain;
+
+public enum LoginType {
+    KAKAO, APPLE
+}
