+++ src/main/java/timify/com/member/domain/MemberStatus.java
@@ -1,4 +1,4 @@
-package timify.com.domain.enums;
+package timify.com.member.domain;
 
 public enum MemberStatus {
     ACTIVE, INACTIVE
